+++ CHANGELOG.md
@@ -28,6 +28,7 @@ v2.0.0-beta.7
  * ons-lazy-repeat: Fixed [[#1236](https://github.com/OnsenUI/OnsenUI/issues/1236), [#1029](https://github.com/OnsenUI/OnsenUI/issues/1029), [#470](https://github.com/OnsenUI/OnsenUI/issues/470)], and [#1035](https://github.com/OnsenUI/OnsenUI/issues/1035).
  * ons-input: Update Angular bindings to work better with checkboxes and radio buttons.
  * ons-switch: Fix so it works inside a <label> element.
+ * ons-ripple: Attach event listeners to parent to avoid blocking cliks.
 
 v2.0.0-beta.6
 ----